Don't use platform.workspace in Reids/Flower resources (#96)
In using this chart outside of the Astronomer platform we noticed that
two places were using the `platform.workspace` that didn't need it.

The broker-url secret is just using it as a label, and that doesn't make
any sense (but also doesn't break anything).

The flower service has it specified in the label selector, so this
service wasn't working if the workspace wasn't provided. For
comparison, the webserver service just has component, release and tier
as its selector.