Allow annotationso to be added to create-user-job (#100)
Some of our users are using Istio, and this is makes the create-user-job not being able to end. In order to fix this we are adding a parameter to be able to add an annotation to that job on a pod level so we can disable de istio sidecar injection.