Support for modifying celery worker deployment strategy (#120)
* Support for modifying the worker deployment strategy

worker template can now accecpt different values for update strategy

Added default set of values for worker update strategy

* Changed worker persistence default to be false

As far as I can tell we are changing this to false in our clouds anyway, possibly because of keda. But I don’t believe this should default to true anyway because of the also large volume size by default. This should be something someone has to explicitly enable

* Added tests for update strategy.