Fix labels set by pod_mutation_hook (#231)
Set the proper values for `kubernetes_executor` and
`kubernetes_pod_operator` labels.

This is critical, especially when using 2.0+, KubernetesExecutor, KPO and
`is_delete_operator_pod=False`, as the `kubernetes_executor` label is used
in the completed pod adoption logic in KubernetesExecutor. With the wrong
labels, the executor tries to adopt the task pod.