Add Nginx auth sidecar -> astronomer ->  airflow deployments  (#221)
This commit adds a mechanism to allow the platform to deploy airflow with auth sidecar .

It operates under the assumption that we will not be able to make any global config changes to OpenShift components (such as the router) and so we have to work around some limitations.

This approach allows astronomer to deploy airflow without  the need of  nginx ingress controller and allow to run with any ingress(ex: openshift routes)


## Background
Under the current design of the platform it is tightly coupled with nginx  ingress controller , and doesn't gives a flexibility to integrate with existing ingress implementation and as a result ending up managing multiple ingress components and additional LoadBalancers. 
## How this approach works

In this new approach all the app deployments includes  a sidecar  which does platform specific stuffs , this enables users to use their own ingress approach to communicate with the application. 


## Drawbacks

individual management of configs, secrets .